requirements.txt
annotated-types==0.6.0
anyio==4.3.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.30.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.110.0
h11==0.14.0
idna==3.6
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
pillow==10.2.0
pydantic==2.6.4
pydantic_core==2.16.3
pypng==0.20220715.0
python-dotenv==1.0.1
pyzbar==0.1.9
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
